requirements.txt
openai-whisper
torch
torchaudio
numpy
ffmpeg-python

scipy
sounddevice
python-dotenv
pyttsx3